+++ src/Spd.Manager.Licence/LicenceAppContract.cs
@@ -54,4 +54,5 @@ public record LicenceAppBase
     public ApplicationTypeCode? ApplicationTypeCode { get; set; }
     public LicenceTermCode? LicenceTermCode { get; set; } //for biz licence term, only 1,2,3 year
     public BizTypeCode? BizTypeCode { get; set; }
+    public ApplicationOriginTypeCode ApplicationOriginTypeCode { get; set; } = ApplicationOriginTypeCode.Portal;
 }
