+++ src/Spd.Resource.Repository/BizLicApplication/Contract.cs
@@ -17,6 +17,7 @@ public record BizLicApplication
     public WorkerLicenceTypeEnum WorkerLicenceTypeCode { get; set; }
     public ApplicationTypeEnum ApplicationTypeCode { get; set; }
     public BizTypeEnum? BizTypeCode { get; set; }
+    public ApplicationOriginTypeCode? ApplicationOriginTypeCode { get; set; } = Repository.ApplicationOriginTypeCode.Portal;
     public string? GivenName { get; set; }
     public string? MiddleName1 { get; set; }
     public string? MiddleName2 { get; set; }
